Support basic basic login.
This rips out the implicit user creation that we were doing before,
and adds slightly more explicit concept of users and auth.

Clients can "login" with a user by invoking the /serve/login path
with an email. This will return the associated user ID or else
create a new user and return that ID.

All other requests must be "authenticated", which for this sample
just means passing a valid userID in the Authorization HTTP header.

Toward https://github.com/rocicorp/replicache/issues/25